// lib/api/generated/schemas/favoriteDto.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * НеМалика API
 * Маркетплейс компьютерной техники — REST API для веб-витрины (Next.js) и Telegram mini-app. Покупать можно без авторизации; вход через Telegram нужен продавцу, администратору и покупателю, который хочет видеть историю просмотров на всех своих устройствах.
 * OpenAPI spec version: 1.0.0
 */
import type { FavoriteDtoState } from './favoriteDtoState';

export interface FavoriteDto {
  /** Идентификатор товара */
  id: number;
  shopId: number;
  shopName: string;
  name: string;
  /** @nullable */
  description: string | null;
  /** Ключи файлов в S3 */
  photos: string[];
  /** numeric приходит строкой */
  price: string;
  state: FavoriteDtoState;
  /** Когда товар добавили в избранное */
  addedAt: string;
}